Adding clean LC8 protein structure
altered via PyMOL 
creates error for topology file conversion 
Fatal error:
Residue 1 named LYS of a molecule in the input file was mapped
to an entry in the topology database, but the atom CG used in
that entry is not found in the input file. Perhaps your atom
and/or residue naming needs to be fixed.